app/requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
supabase
pydantic[email]